refactor: Improve folder structure
